refactor(nes-cpu): split impl Cpu into focused submodules (#2827)
Distribute the single ~2.3K-line impl Cpu block across cohesive
submodules under src/nes/cpu/cpu/: state, interrupts, dma, timing, bus,
alu, execute. mod.rs retains the type definitions and new(). Methods
called across module boundaries are elevated to pub(super); no item is
exposed beyond the cpu module subtree, so the external API is unchanged.

Pure code move: no behavior change. All 680 CPU tests pass; every impl
file is now under 1,500 lines (largest: execute.rs at 900).
